+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Semantic versioning
 Version latest (0.18.0rc?, 2019-12-05?)
 
 - ...
+- enh: use pytest parametrize for 2 tests and reduce boilerplate code
 
 
 Version 0.19.2 (2019-12-19)